11/1

+++ ft_putadrs.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putadrs.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ichinoseyuuki <ichinoseyuuki@student.42    +#+  +:+       +#+        */
+/*   By: yichinos <yichinos@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/31 14:41:17 by yichinos          #+#    #+#             */
-/*   Updated: 2022/10/31 21:22:45 by ichinoseyuu      ###   ########.fr       */
+/*   Updated: 2022/11/01 14:35:32 by yichinos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@ int	ft_putadrs(uintptr_t a)
 	count += write(1, "0x", 2);
 	if (a == 0)
 		count += write(1, 0, 1);
-	count += ft_putnbr_small_sixteen(a);
+	count += ft_putnbr_small_sixteen((unsigned long int)a);
 	return (count);
 }
